+++ Codebase/Frontend/CalendarUI/index.html
@@ -14,7 +14,6 @@
         <script src="https://cdn.jsdelivr.net/npm/@fullcalendar/bootstrap5@6.1.15/index.global.min.js"></script>
         <!-- Import local CSS & JS files -->
         <link rel="stylesheet" href="style.css"/>
-        <script type="module" src="script_firebase.js"></script>
         <script type="module" src="script.js"></script>
     </head>
     <body>
