+++ Codebase/Frontend/CalendarUI/script.js
@@ -1,42 +1,160 @@
-// Fetch data from Firebase in the form of JSON
-async function getJSON() {
-  const url = 'https://clarity-295d8-default-rtdb.firebaseio.com/eventDB.json';
-  try {
-    const response = await fetch(url); // Make HTTP request to URL, await the response
-    if (!response.ok) { // If the response status isn't OK (200), throw an error containing the actual status
-      throw new Error('HTTP Status ' + response.status + ', ' + response.statusText);
+import { initializeApp } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-app.js";
+import { getDatabase, set, ref, child, get, push, onValue, update } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-database.js";
+
+// Firebase configuration
+const firebaseConfig = {
+    apiKey: "AIzaSyAAxXs88pJUfCoeotb0C8gfTGxvltpPBz8",
+    authDomain: "clarity-295d8.firebaseapp.com",
+    databaseURL: "https://clarity-295d8-default-rtdb.firebaseio.com",
+    projectId: "clarity-295d8",
+    storageBucket: "clarity-295d8.appspot.com",
+    messagingSenderId: "117186684063",
+    appId: "1:117186684063:web:a0a70113604e5c07ed2eaa" 
+};
+
+// Initialize Firebase
+const app = initializeApp(firebaseConfig);
+const db = getDatabase(app);
+
+let currentTask = null;
+let taskNameInput, attributeInput, dateDueInput, prioritySelect, modalOverlay;
+
+// Wait for Modal input fields to load
+document.addEventListener("DOMContentLoaded", () => {
+    taskNameInput = document.getElementById('taskName');
+    attributeInput = document.getElementById('attribute');
+    dateDueInput = document.getElementById('dateDue');
+    prioritySelect = document.getElementById('priority');
+    modalOverlay = document.getElementById('modalOverlay');
+});
+
+// Open the Modal (by unhiding it) & clearing all entry fields
+window.openModal = function(taskElement = null) {
+    currentTask = taskElement;
+    if (taskElement) {
+        const taskId = taskElement.getAttribute('data-id');
+        const taskRef = ref(db, `eventDB/${taskId}`);
+
+        onValue(taskRef, (snapshot) => {
+            const childData = snapshot.val();
+            if (childData) {
+                taskNameInput.value = childData.title;
+                attributeInput.value = childData.attribute || "";
+                dateDueInput.value = childData.dateDue || "";
+                prioritySelect.value = childData.priority || "";
+            }
+        }, { onlyOnce: true });
+    } else {
+        // Clear modal fields for new task
+        taskNameInput.value = "";
+        attributeInput.value = "";
+        dateDueInput.value = "";
+        prioritySelect.value = "";
     }
-    return response.json(); // Upon fetch success, parse the response data as a JSON object
-  } catch (error) { // Send error message to the browser console
-    console.error(error.message);
-    return JSON.parse(""); // Return an empty JSON object upon failure
-  }
+    modalOverlay.style.display = 'flex';
 }
 
-// Parse JSON into an array of Calendar events
-function parseJSON(input) { // TODO: Add optional fields to the parser + some error checking
-  //console.log(input); // For debugging purposes, disable before release
-  var events = new Array();
-  for (var key in input) { events.push({
-    title : (input[key])["title"],
-    start : (input[key])["start"],
-    end : (input[key])["end"],
-    description : (input[key])["attribute"]});
-  }
-  return events;
+// Close the Modal (by hiding it)
+window.closeModal = function() {
+    modalOverlay.style.display = 'none';
+    currentTask = null;
 }
 
-;(async () => { // IIFE used in order to use await at the top level
-  const JSON_data = await getJSON(); // Await response before storing JSON data
-  var event_data = parseJSON(JSON_data); // Parse JSON data for events
-  //console.log(event_data); // For debugging purposes, disable before release
+// Save task, update Firebase, & close the Modal
+window.saveTask = function() {
+    const updatedName = taskNameInput.value.trim();
+    const updatedAttribute = attributeInput.value.trim();
+    const updatedDateDue = dateDueInput.value;
+    const updatedPriority = prioritySelect.value;
+
+    if (!updatedName) {
+        alert("Task name is required.");
+        return;
+    }
+    if (currentTask) {
+        // Update existing task
+        const taskId = currentTask.getAttribute('data-id');
+        const taskRef = ref(db, `eventDB/${taskId}`);
+        update(taskRef, {
+            title: updatedName,
+            attribute: updatedAttribute,
+            dateDue: updatedDateDue,
+            priority: parseInt(updatedPriority)
+        }).then(() => {
+            console.log("Task updated successfully in Firebase.");
+        }).catch((error) => {
+            console.error("Error updating task:", error);
+        });
+        currentTask.textContent = `${updatedName} (Priority: ${updatedPriority})`;
+    } else {
+        // Create new task
+        const start = updatedDateDue + 'T09:00:00';
+        const end = updatedDateDue + 'T17:00:00';
+        const dateAdded = new Date().toISOString();
+        const eventData = {
+            title: updatedName,
+            attribute: updatedAttribute,
+            dateAdded,
+            start,
+            end,
+            priority: parseInt(updatedPriority)
+        };
+        const eventRef = push(ref(db, 'eventDB/'));
+        set(eventRef, eventData)
+            .then(() => {
+                console.log("Event saved successfully.");
+            })
+            .catch((error) => {
+                console.error("Error saving event:", error);
+            });
+    }
+    closeModal();
+}
+
+// Main loader for the Calendar
+document.addEventListener('DOMContentLoaded', function() {
+  var event_data = new Array();
+  
+  const dbRef = ref(getDatabase());
+  get(child(dbRef, "eventDB/")).then((snapshot) => {
+    if (snapshot.exists()) {
+      const raw_data = snapshot.val();
+      for (var key in raw_data) { 
+        var temp_start = (raw_data[key])["start"];
+        var temp_end = (raw_data[key])["end"];
+        var temp_desc = (raw_data[key])["attribute"];
+        if (temp_start.startsWith('T') || temp_start == "") {
+          temp_start = (new Date().toISOString()).substring(0, 19);
+        } 
+        if (temp_end.startsWith('T') || temp_end == "") {
+          temp_end = (new Date().toISOString()).substring(0, 19);
+        } else if (!temp_end.includes('T')) {
+          temp_end += "T23:59:00";
+        }
+        if (temp_desc == "" || temp_desc == " ") {
+          temp_desc = "No description.";
+        }
+        event_data.push({
+          title : (raw_data[key])["title"],
+          start : temp_start,
+          end : temp_end,
+          description : temp_desc
+        });
+      }
+    } else {
+      console.log("No data available");
+    }
+  }).catch((error) => {
+    console.error(error);
+  });
+
+  console.log(event_data); // For debugging purposes, disable before release
 
   var calendarEl = document.getElementById('calendar'); // Get 'calendar' object from page
 
   var calendar = new FullCalendar.Calendar(calendarEl, { // Instance new calendar object
     themeSystem: 'bootstrap5',
     initialView: 'dayGridMonth',
-    editable: true,
     dayMaxEvents: true,
     navLinks: true,
     headerToolbar: {
@@ -63,8 +181,9 @@ function parseJSON(input) { // TODO: Add optional fields to the parser + some er
         container: 'body'
       });
     },
-    events: event_data, // Use event_data array
-    eventColor: '#d9ae89'
+    eventColor: '#d9ae89',
+    eventSources: [event_data]
   });
   calendar.render(); // Display the calendar on the page
-})();
+  console.log(calendar.getEventSources()); // For debugging purposes, disable before release
+});
